# src/public-website/contracts/openapi/admin-api.yaml
openapi: 3.1.0
info:
  title: International Center Admin API
  description: Administrative API for International Center content management system
  version: 1.0.0
  contact:
    name: International Center API Team
servers:
  - url: https://admin-api.axiomcloud.dev/api/v1
    description: Production admin server
  - url: http://localhost:4001/api/v1
    description: Development admin server
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /inquiries:
    get:
      summary: Get all inquiries
      operationId: getInquiries
      tags:
        - Inquiries Management
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - name: inquiry_type
          in: query
          schema:
            type: string
            enum: [media, business, donation, volunteer]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, closed]
      responses:
        '200':
          description: List of inquiries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inquiry'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
  /inquiries/{id}:
    get:
      summary: Get inquiry by ID
      operationId: getInquiryById
      tags:
        - Inquiries Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Inquiry details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Inquiry'

    put:
      summary: Update inquiry status
      operationId: updateInquiryStatus
      tags:
        - Inquiries Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, in_progress, completed, closed]
                notes:
                  type: string
                  nullable: true
                assigned_to:
                  type: string
                  format: uuid
                  nullable: true
              required:
                - status
      responses:
        '200':
          $ref: '#/components/responses/UpdatedResponse'

  # Analytics endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      $ref: './components/parameters/pagination.yaml#/PageParam'
    LimitParam:
      $ref: './components/parameters/pagination.yaml#/LimitParam'
    SearchParam:
      name: search
      in: query
      description: Search query string for full-text search
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 255
      example: "cancer research"
...
  schemas:
...
    PaginationInfo:
      $ref: './public-api.yaml#/components/schemas/PaginationInfo'

    # Reference existing domain schemas
...
    Inquiry:
      type: object
      properties:
        inquiry_id:
          type: string
          format: uuid
        inquiry_type:
          type: string
          enum: [media, business, donation, volunteer]
        status:
          type: string
          enum: [pending, in_progress, completed, closed]
        submitter_name:
          type: string
        submitter_email:
          type: string
          format: email
        subject:
          type: string
        message:
          type: string
        submitted_on:
          type: string
          format: date-time
        assigned_to:
          type: string
          format: uuid
          nullable: true
        notes:
          type: string
          nullable: true
        last_updated:
          type: string
          format: date-time
          nullable: true
      required:
        - inquiry_id
        - inquiry_type
        - status
        - submitter_name
        - submitter_email
        - subject
        - message
        - submitted_on
...
  responses:
...
    UpdatedResponse:
      $ref: './components/responses/success.yaml#/UpdatedResponse'
...